Improve code comments and add lint check on npm start script
Add more comments to the code following the JSDOC specs.
We need later to be able to use JSDOC to generate documentation
of this project so it is important to ensure all code comments follow
the JSDOC specification.

Add lint check prior to npm start script.
This helps to catch and solve lint error whenever they
appear when the app is running in development mode compared
to the previous style of tracing lint error independently of the
start script so developer had to stop the running app and run
the lint script to fix lint error. It was annoying and time

Closes: #51, #52, #50, #44, #43, #42, #40, #39, #37