Fix page scroll in stakeholder page
Removing page scroll in stakeholder page. Stakeholder page is
composed of three columns with content in each column. When you
were scrolling the column content the whole page was scrolling as
well, something which was not good UI experience so I removed
the page scroll in favour of column scroll