Plans: Refactor Plan folder structure
This change name of components so that they make more sense, i.e now
actions have been renamed to activities. This is a convention we agreed
that activity instance will be called an action but during planning it
will be called activity.

This also Add activity form and a component to display stakeholders
details in activity detailed view.

Also this improve documentation in the components