Fix stakeholder font type and loading state UI
Stakeholder fonts were default to System UI font while EMIS
is using lato. This fix ensure the font used across is of lato
family.

The loading state was not logical since it was only showing
in the single component despite the fact that all components
under stakeholder were dependant to the loading of the data.
This fix hide all components that are not useful when the loading
is not finished.